requirements.txt
chardet==4.0.0
iribaker==0.2
isodate==0.6.1
Jinja2==3.0.3
Js2Py==0.71
pyjsparser==2.7.1
pytz==2021.3
PyYAML==6.0
rdflib==6.0.2
rdflib-jsonld==0.6.2
rfc3987==1.3.8
tzlocal==4.1
unicodecsv==0.14.1
Werkzeug==2.0.2
